Moved UserRegisterController to controllers package.
Made new presenters package and moved all presenters there (to break up views package).
