+++ src/main/java/Main.java
@@ -18,6 +18,7 @@ import message_search_use_case.MessageSearchInputBoundary;
 import message_search_use_case.MessageSearchInteractor;
 import message_search_use_case.MessageSearchOutputBoundary;
 import org.jetbrains.annotations.NotNull;
+import presenters.*;
 import profile_customization_use_case.CustomizationGateway;
 import profile_customization_use_case.CustomizationInputBoundary;
 import profile_customization_use_case.CustomizationInteractor;
