+++ src/main/java/controllers/UserRegisterController.java
@@ -1,4 +1,4 @@
-package views;
+package controllers;
 
 
 import user_register_use_case.CreationData;
