+++ src/main/java/presenters/MessageEditPresenter.java
@@ -1,4 +1,4 @@
-package views;
+package presenters;
 
 import message_edit_delete_use_case.MessageEditFailed;
 import message_edit_delete_use_case.MessageEditResponse;
