+++ src/test/java/message_edit_delete_use_case/MessageDeleteInteractorTest.java
@@ -4,14 +4,13 @@ import com.google.cloud.firestore.DocumentReference;
 import com.google.cloud.firestore.Firestore;
 import com.google.firebase.cloud.FirestoreClient;
 import gateways.MessageDeleteFirebaseSystem;
-import message_search_use_case.MessageSearchResponse;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import services.DBInitializer;
-import views.MessageDeletePresenter;
+import presenters.MessageDeletePresenter;
 
 import java.io.FileNotFoundException;
 import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.ExecutionException;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -28,7 +27,7 @@ class MessageDeleteInteractorTest {
         Firestore dbFirestore = FirestoreClient.getFirestore();
         interactor.messageDelete(new MessageDeleteData(5, 4));
         DocumentReference chatref = dbFirestore.collection("chats").document("id"+4);
-        assertEquals(0, ( (List<DocumentReference>)chatref.get().get().getData().get("messages")).size());
+        assertEquals(0, ( (List<DocumentReference>) Objects.requireNonNull(chatref.get().get().getData()).get("messages")).size());
 
     }
 }
