+++ src/test/java/message_search_use_case/MessageSearchInteractorTest.java
@@ -4,7 +4,7 @@ import gateways.MessageSearchFirebaseSystem;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import services.DBInitializer;
-import views.MessageSearchPresenter;
+import presenters.MessageSearchPresenter;
 
 import java.io.FileNotFoundException;
