+++ src/test/java/views/MessageEditControllerTest.java
@@ -8,6 +8,7 @@ import gateways.MessageEditFirebaseSystem;
 import message_edit_delete_use_case.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import presenters.MessageEditPresenter;
 import services.DBInitializer;
 
 import java.io.FileNotFoundException;
