+++ src/test/java/views/MessageSearchPresenterTest.java
@@ -6,6 +6,7 @@ import message_search_use_case.MessageSearchFailed;
 import message_search_use_case.MessageSearchResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import presenters.MessageSearchPresenter;
 
 import java.util.ArrayList;
 import java.util.Date;
